Expense app with user authentication

+++ controllers/ExpenseTracker/expenses.js
@@ -1,20 +1,22 @@
 const Expenses = require('../../models/ExpenseTracker/expenses');
 
-exports.getAll = async (req, res, next) => {
+exports.getAllExpenses = async (req, res, next) => {
   try{
 
-      const AllUsers = await Expenses.findAll();
-      res.status(201).json( {message: "Successfully extracted", allExpenseData:AllUsers}); 
+      const AllExpenses = await req.user.getExpenses();
+      // console.log(AllExpenses);
+      res.status(201).json( {message: "Successfully extracted", allExpenseData:AllExpenses, success:true}); 
   } 
   catch(err){
-    res.status(500).json({message: "Error while fetching already present expenses"})
+    console.log(err);
+    res.status(500).json({message: "Error while fetching already present expenses", success:false})
   }
 }
 
 exports.postExpense = async (req, res, next) => {
   try{
     const {amount, category, description} = req.body;
-    console.log(amount, category, description)
+    const idUser = req.user.id;
     if(!amount | !category | !description){
       return res.status(500).json({message: 'All fields are mandatory'})
     }
@@ -22,7 +24,8 @@ exports.postExpense = async (req, res, next) => {
       const data = await Expenses.create({
         amountDB: amount,
         categoryDB: category,
-        descriptionDB: description
+        descriptionDB: description,
+        userId: idUser
       });
       res.status(201).json({newExpenseData: data});
     }
@@ -34,11 +37,12 @@ exports.postExpense = async (req, res, next) => {
 
 exports.delExpense = async (req, res, next) => {
     const deleteId = req.params.delId;
+    const idUser = req.user.id;
     try{
-        const delUser =   await Expenses.destroy( { where: { id:deleteId } });
-        res.status(201).json({delUserfromDB: delUser })
+      const delUser = await Expenses.destroy( { where: { id:deleteId, userId:idUser } });
+      res.status(201).json({delUserfromDB: delUser })
     }
     catch(err){
-        res.status(500).json({error: err})
+      res.status(500).json({error: err})
     }
 }
